Changed protocol to a more reliable protocol

+++ CNCController/CommResult.cs
@@ -6,7 +6,7 @@ namespace CNCController
 {
     public class CommResult
     {
-        private CommResult(ulong id, Task send, Task confirmed, Task completed)
+        private CommResult(uint id, Task send, Task confirmed, Task completed)
         {
             this.Id = id;
             this.Send = send;
@@ -14,12 +14,12 @@ namespace CNCController
             this.Completed = completed;
         }
 
-        public ulong Id { get; }
+        public uint Id { get; }
         public Task Send { get; }
         public Task Completed { get; }
         public Task Confirmed { get; }
 
-        internal static CommResult Create(ulong commandId, Task send, TaskCompletionSource<ulong> tcsAck, TaskCompletionSource<ulong> tcsCmp, CancellationToken cancel)
+        internal static CommResult Create(uint commandId, Task send, TaskCompletionSource<uint> tcsAck, TaskCompletionSource<uint> tcsCmp, CancellationToken cancel)
         {
             send.ContinueWith(t => cascadeCancelFail(t, tcsAck));
             cancel.Register(() => tcsAck.TrySetCanceled());
@@ -31,7 +31,7 @@ namespace CNCController
             return new CommResult(commandId, send, tcsAck.Task, tcsCmp.Task);
         }
 
-        private static void cascadeCancelFail(Task t, TaskCompletionSource<ulong> tcsAck)
+        private static void cascadeCancelFail(Task t, TaskCompletionSource<uint> tcsAck)
         {
             if (t.IsCanceled)
                 tcsAck.TrySetCanceled();
